Consolidate from logic and image version resolution in general
