+++ aoc_cli/commands/init.py
@@ -25,6 +25,7 @@ from aoc_cli.services.ingress import IngressService
 from aoc_cli.services.influxdb import InfluxDBService
 from aoc_cli.services.keycloak import KeycloakConfig, KeycloakService
 from aoc_cli.utils.secrets import generate_secret
+from aoc_cli.utils.images import resolve_images, replace_docker_compose_services_versions
 
 
 def _is_in_automation_center_repo() -> bool:
@@ -53,25 +54,6 @@ def _is_in_automation_center_repo() -> bool:
         return False
 
 
-def _find_django_directory() -> Path:
-    """Find the django directory using absolute path from git repo root."""
-    try:
-        # Get the git root directory
-        result = subprocess.run(
-            ["git", "rev-parse", "--show-toplevel"],
-            capture_output=True,
-            text=True,
-            check=True
-        )
-        git_root = Path(result.stdout.strip())
-        django_dir = git_root / "django"
-        
-        if not django_dir.exists():
-            raise FileNotFoundError("Django directory not found in git repository")
-            
-        return django_dir
-    except (subprocess.CalledProcessError, FileNotFoundError) as e:
-        raise RuntimeError(f"Could not find django directory: {e}")
 
 
 @click.command()
@@ -123,19 +105,13 @@ def _find_django_directory() -> Path:
     type=click.Path(exists=True, file_okay=False, dir_okay=True),
     help="The directory where the certificates are located",
 )
+@click.option("--from", "from_url", type=str, help="URL endpoint returning versions for update")
 @click.pass_context
 def init(
-    ctx,
-    output_dir: Path,
-    env_file,
-    overwrite,
-    dev,
-    continue_from_config,
-    mkcerts,
-    certs_dir,
+    *args, 
     **kwargs,
 ):
-    asyncio.run(_init_async(ctx, output_dir, env_file, overwrite, dev, continue_from_config, mkcerts, certs_dir, **kwargs))
+    asyncio.run(_init_async(*args, **kwargs))
 
 
 async def _init_async(
@@ -147,6 +123,7 @@ async def _init_async(
     continue_from_config,
     mkcerts,
     certs_dir,
+    from_url,
     **kwargs,
 ):
     """Initialize the Automation Operations Center (AOC)."""
@@ -207,6 +184,7 @@ async def _init_async(
             keycloak_smtp_port=kwargs["keycloak_smtp_port"],
             mkcerts=mkcerts,
             certs_dir=certs_dir,
+            from_url=from_url,
         )
         
         # Save config to YAML file
@@ -261,6 +239,7 @@ async def _init_async(
         keycloak_smtp_port=kwargs.get("keycloak_smtp_port"),
         mkcerts=mkcerts,
         certs_dir=certs_dir,
+        from_url=from_url,
     )
 
     await execute_init(init_config)
@@ -406,104 +385,6 @@ async def execute_init(config: InitConfig, continue_from_config: bool = False) -
 
 def create_aoc_directory(config: InitConfig) -> None:
     config.aoc_dir.mkdir(parents=True, exist_ok=True)
-
-
-def resolve_images(config: InitConfig) -> None:
-    """Ensure config image fields are full image references with tags.
-
-    If a field is unset, query Docker Hub and pick the latest tag matching
-    the expected pattern, then set the config field accordingly.
-    """
-    def latest_tag_from_docker_hub(url: str, pattern: str = r"\d{4}-\d+-git-[a-fA-F0-9]+$") -> str:
-        try:
-            resp = requests.get(url, timeout=15)
-            if resp.status_code == 200:
-                results = resp.json().get("results", [])
-                for version in results:
-                    name = version.get("name")
-                    if name and re.match(pattern, name):
-                        return name
-        except Exception:
-            pass
-        return "latest"
-
-    # bitswan-backend
-    if not config.aoc_be_image:
-        tag = latest_tag_from_docker_hub(
-            "https://hub.docker.com/v2/repositories/bitswan/bitswan-backend/tags/"
-        )
-        config.aoc_be_image = f"bitswan/bitswan-backend:{tag}"
-
-    # automation-operations-centre (AOC)
-    if not config.aoc_image:
-        tag = latest_tag_from_docker_hub(
-            "https://hub.docker.com/v2/repositories/bitswan/automation-operations-centre/tags/"
-        )
-        config.aoc_image = f"bitswan/automation-operations-centre:{tag}"
-
-    # profile-manager
-    if not config.profile_manager_image:
-        tag = latest_tag_from_docker_hub(
-            "https://hub.docker.com/v2/repositories/bitswan/profile-manager/tags/"
-        )
-        config.profile_manager_image = f"bitswan/profile-manager:{tag}"
-
-    # keycloak
-    if not config.keycloak_image:
-        tag = latest_tag_from_docker_hub(
-            "https://hub.docker.com/v2/repositories/bitswan/bitswan-keycloak/tags/"
-        )
-        config.keycloak_image = f"bitswan/bitswan-keycloak:{tag}"
-
-
-def replace_docker_compose_services_versions(config: InitConfig) -> None:
-    click.echo("Finding latest versions for AOC services")
-
-    # Read the docker-compose.yml file using yaml
-    with open(config.aoc_dir / "docker-compose.yml", "r") as f:
-        docker_compose = yaml.safe_load(f)
-
-    # Images are expected to be resolved already in config
-    service_to_image = {
-        "bitswan-backend": config.aoc_be_image,
-        "aoc": config.aoc_image,
-        "profile-manager": config.profile_manager_image,
-        "keycloak": config.keycloak_image,
-    }
-
-    for service_name, image in service_to_image.items():
-        if image:
-            docker_compose["services"][service_name]["image"] = image
-
-    # Add dev mode configuration for bitswan-backend service
-    if config.env == Environment.DEV:
-        try:
-            django_dir = _find_django_directory()
-            click.echo(f"Adding dev mode volume mapping: {django_dir}:/app:rwz")
-            
-            # Ensure the bitswan-backend service exists
-            if "bitswan-backend" not in docker_compose["services"]:
-                docker_compose["services"]["bitswan-backend"] = {}
-            
-            # Add volume mapping for dev mode
-            if "volumes" not in docker_compose["services"]["bitswan-backend"]:
-                docker_compose["services"]["bitswan-backend"]["volumes"] = []
-            
-            # Add the django directory volume mapping
-            volume_mapping = f"{django_dir}:/app:rwz"
-            if volume_mapping not in docker_compose["services"]["bitswan-backend"]["volumes"]:
-                docker_compose["services"]["bitswan-backend"]["volumes"].append(volume_mapping)
-            
-            # Ensure command is set to /start for dev mode
-            docker_compose["services"]["bitswan-backend"]["command"] = "/dev-command"
-            
-        except Exception as e:
-            click.echo(f"Warning: Could not configure dev mode volumes: {e}")
-            click.echo("Continuing without dev mode volume mapping...")
-
-    # Write the updated docker-compose.yml file
-    with open(config.aoc_dir / "docker-compose.yml", "w") as f:
-        yaml.dump(docker_compose, f)
 
 
 def setup_keycloak(config: InitConfig) -> None:
